+++ autoprotocol/instruction.py
@@ -1140,3 +1140,22 @@ class MeasureMass(Instruction):
                      "object": refs,
                      "dataref": dataref}
         super(MeasureMass, self).__init__(json_dict)
+
+
+class MeasureVolume(Instruction):
+    """
+    Measure the mass of containers
+
+    Parameters
+    ----------
+    refs : list of containers
+        list of containers
+    dataref: str
+        Name of the data for the measurement
+    """
+
+    def __init__(self, wells, dataref):
+        json_dict = {"op": "measure_volume",
+                     "object": wells,
+                     "dataref": dataref}
+        super(MeasureVolume, self).__init__(json_dict)